action.tsに統合
